+++ src/lib/algorithm/cipher/mode/BlockCipherMode.ts
@@ -1,4 +1,5 @@
 import type {BlockCipher} from "../BlockCipher";
+import type {Word32Array} from "../../../Word32Array";
 
 export interface BlockCipherModeProps {
   cipher: BlockCipher;
@@ -23,8 +24,15 @@ export class BlockCipherMode {
   /**
    * @abstract
    */
-  public processBlock(words: number[], offset: number){
+  public processBlock(words: number[], offset: number) {
     return;
+  }
+  
+  /**
+   * @abstract
+   */
+  public generateAuthTag(cipherText: Word32Array): Word32Array|undefined {
+    return undefined;
   }
   
   /**