+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.0] - 2021-04-07
+### Added
+- Added GCM block cipher mode. (Galois Counter Mode)
+- Added GMAC (Galois Message Authentication Code)
+
 ## [0.1.0] - 2021-03-30
 ### Added
 - Added CLI executable  
@@ -40,7 +45,8 @@
 ## [0.0.0] - 2021-03-21
 Initial release.
 
-<!-- [Unreleased]: https://github.com/Hinaser/jscrypto/compare/v0.1.0...v0.1.1 -->
+<!-- [Unreleased]: https://github.com/Hinaser/jscrypto/compare/v0.1.0...v0.2.0 -->
+[0.2.0]: https://github.com/Hinaser/jscrypto/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/Hinaser/jscrypto/compare/v0.0.2...v0.1.0
 [0.0.2]: https://github.com/Hinaser/jscrypto/releases/tag/v0.0.2
 [0.0.1]: https://github.com/Hinaser/jscrypto/releases/tag/v0.0.1